Move ProgramError into sdk/
