+++ src/bank.rs
@@ -19,12 +19,12 @@ use native_loader;
 use payment_plan::Payment;
 use poh_recorder::PohRecorder;
 use poh_service::NUM_TICKS_PER_SECOND;
-use program::ProgramError;
 use rayon::prelude::*;
 use rpc::RpcSignatureStatus;
 use runtime::{self, RuntimeError};
 use solana_sdk::account::Account;
 use solana_sdk::hash::{hash, Hash};
+use solana_sdk::native_program::ProgramError;
 use solana_sdk::pubkey::Pubkey;
 use solana_sdk::signature::Keypair;
 use solana_sdk::signature::Signature;
