+++ src/budget_program.rs
@@ -4,8 +4,8 @@ use budget_expr::BudgetExpr;
 use budget_instruction::Instruction;
 use chrono::prelude::{DateTime, Utc};
 use payment_plan::Witness;
-use program::ProgramError;
 use solana_sdk::account::Account;
+use solana_sdk::native_program::ProgramError;
 use solana_sdk::pubkey::Pubkey;
 use solana_sdk::transaction::Transaction;
 use std::io;
